backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
anthropic>=0.18.0
pydantic>=2.5.0
PyMuPDF>=1.23.0
httpx>=0.26.0
firecrawl-py>=1.0.0
